examples/deployment/modal-example/requirements.txt
python-dotenv==1.0.1
modal==0.65.48
pipecat-ai[daily,silero,cartesia,openai,whisper,deepgram]
fastapi==0.115.4
aiohttp
deepgram-sdk